extend some hashcode combiner implementations
Add implementations for the custom multiply-add and murmur-3-32
hash-code combiners that separate getting the hash-codes from
mixing them.

Theoretically this should improve performance because the mixing
steps themselves are smaller than the full .GetHashCode() machinery,
and the kind of operations should be able to better utilize the
CPU's resources when they are not interrupted.

Practically, the multiply-add combiner gets faster on my notebook,
the murmur3-32 combiner not.